refactor: image optimization
